Extract imagetype-to-keytype statements into a utility function and move tvdb specific utils to separate class
