+++ MediaBrowser.Controller/Library/TVUtils.cs
@@ -8,26 +8,6 @@ namespace MediaBrowser.Controller.Library
     public static class TVUtils
     {
         /// <summary>
-        /// The TVDB API key
-        /// </summary>
-        public static readonly string TvdbApiKey = "OG4V3YJ3FAP7FP2K";
-        public static readonly string TvdbBaseUrl = "https://www.thetvdb.com/";
-        /// <summary>
-        /// The banner URL
-        /// </summary>
-        public static readonly string BannerUrl = TvdbBaseUrl + "banners/";
-
-        public static string NormalizeLanguage(string language)
-        {
-            if (string.IsNullOrWhiteSpace(language))
-            {
-                return null;
-            }
-
-            // pt-br is just pt to tvdb
-            return language.Split('-')[0].ToLowerInvariant();
-        }
-        /// <summary>
         /// Gets the air days.
         /// </summary>
         /// <param name="day">The day.</param>
